Expand User to fit spec, finish endpoints

+++ src/main/java/com/booleanuk/api/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.booleanuk.api.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import com.booleanuk.api.model.User;
 import com.booleanuk.api.model.UserDTO;
@@ -8,6 +9,7 @@ import com.booleanuk.api.repository.UserRepository;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("users")
@@ -31,29 +33,36 @@ public class UserController {
   @ResponseStatus(HttpStatus.CREATED)
   @PostMapping
   public User create(@RequestBody UserDTO request) {
-    User user = new User(request.email(), request.firstName());
+    User user = new User(request.email(), request.firstName(), request.lastName(), request.username(), request.phone());
     return this.repository.save(user);
+  }
+
+  @ResponseStatus(HttpStatus.CREATED)
+  @PutMapping("{id}")
+  public User put(@PathVariable("id") Integer id, @RequestBody UserDTO request) throws ResponseStatusException {
+    Optional<User> maybeUserToUpdate = this.repository.findById(id);
+    if (maybeUserToUpdate.isEmpty())
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
+    User userToUpdate = maybeUserToUpdate.get();
+    userToUpdate.setEmail(request.email());
+    userToUpdate.setFirstName(request.firstName());
+    userToUpdate.setLastName(request.lastName());
+    userToUpdate.setUsername(request.username());
+    userToUpdate.setPhone(request.phone());
+
+    return this.repository.save(userToUpdate);
   }
 
   @ResponseStatus(HttpStatus.OK)
   @DeleteMapping("{id}")
   public User delete(@PathVariable("id") Integer id) {
-    User deletedRef = this.repository.getReferenceById(id);
-    User deletedCopy = new User(deletedRef.getId(), deletedRef.getEmail(), deletedRef.getFirstName(),
-        deletedRef.getIsActive());
+    Optional<User> maybeUserToUpdate = this.repository.findById(id);
+    if (maybeUserToUpdate.isEmpty())
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
     this.repository.deleteById(id);
 
-    return deletedCopy;
-  }
-
-  @ResponseStatus(HttpStatus.CREATED)
-  @PutMapping("{id}")
-  public User put(@PathVariable("id") Integer id, @RequestBody UserDTO request) {
-    User toUpdateRef = this.repository.getReferenceById(id);
-    toUpdateRef.setFirstName(request.firstName());
-    toUpdateRef.setEmail(request.email());
-
-    return new User(toUpdateRef.getId(), toUpdateRef.getEmail(), toUpdateRef.getFirstName(),
-        toUpdateRef.getIsActive());
+    return maybeUserToUpdate.get();
   }
 }
